refactor: 토큰 패키지를 web 패키지 내부로 이동
