+++ src/main/java/roomescape/service/MemberService.java
@@ -3,7 +3,7 @@ package roomescape.service;
 import org.springframework.stereotype.Service;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
-import roomescape.domain.token.TokenProvider;
+import roomescape.web.api.token.TokenProvider;
 
 import java.util.List;
