+++ src/main/java/roomescape/web/api/token/TokenParser.java
@@ -1,4 +1,4 @@
-package roomescape.domain.token;
+package roomescape.web.api.token;
 
 import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Value;
